+++ petweb/account/urls.py
@@ -1,9 +1,9 @@
 from django.conf.urls import url
 
-from . import views
+from . import apis
 
 urlpatterns = [
-    url(r'^login/$', views.Login.as_view()),
-    url(r'^signup/$', views.Signup.as_view()),
-    url(r'^detail/(?P<pk>\d+)/$', views.UserDetail.as_view()),
+    url(r'^login/$', apis.Login.as_view()),
+    url(r'^signup/$', apis.Signup.as_view()),
+    url(r'^detail/(?P<pk>\d+)/$', apis.UserDetail.as_view()),
 ]
